lc_warehouse_mock_server/requirements.txt
Flask==3.0.0
PyJWT==2.8.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
